% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrambot.R
\name{diagrambot}
\alias{diagrambot}
\title{Interactive Shiny app that generates Mermaid and Graphviz diagrams from voice commands using
GPT-4o Realtime.}
\usage{
diagrambot(debug = FALSE)
}
\arguments{
\item{debug}{Logical; if TRUE, enables debug mode in shinyrealtime for
(extremely) verbose logging.}
}
\value{
A Shiny app object. If this function is called from the R console,
the app will launch automatically.
}
\description{
Interactive Shiny app that generates Mermaid and Graphviz diagrams from voice commands using
GPT-4o Realtime.
}
