% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrambot.R
\name{render_graphviz}
\alias{render_graphviz}
\title{Render a Graphviz diagram}
\usage{
render_graphviz(code)
}
\arguments{
\item{code}{The Graphviz DOT code}
}
\value{
HTML tags for rendering the diagram
}
\description{
Render a Graphviz diagram
}
\keyword{internal}
