% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrambot.R
\name{render_mermaid}
\alias{render_mermaid}
\title{Render a Mermaid diagram}
\usage{
render_mermaid(code)
}
\arguments{
\item{code}{The Mermaid diagram code}
}
\value{
HTML tags for rendering the diagram
}
\description{
Render a Mermaid diagram
}
\keyword{internal}
