% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrambot.R
\name{setup_copy_button_handler}
\alias{setup_copy_button_handler}
\title{Setup copy to clipboard button handler}
\usage{
setup_copy_button_handler(input, session, last_code)
}
\arguments{
\item{input}{The Shiny input object}

\item{session}{The Shiny session object}

\item{last_code}{Reactive value containing the diagram code}
}
\description{
Setup copy to clipboard button handler
}
\keyword{internal}
